+++ zpa/services/policysetcontrollerv2/policysetcontrollerv2.go
@@ -7,6 +7,7 @@ import (
 	"net/http"
 	"sort"
 	"strings"
+	"sync"
 
 	"github.com/zscaler/zscaler-sdk-go/v2/zpa/services"
 	"github.com/zscaler/zscaler-sdk-go/v2/zpa/services/appconnectorgroup"
@@ -19,6 +20,8 @@ const (
 	mgmtConfigV1 = "/mgmtconfig/v1/admin/customers/"
 	mgmtConfigV2 = "/mgmtconfig/v2/admin/customers/"
 )
+
+var ruleMutex sync.Mutex
 
 type PolicySet struct {
 	CreationTime    string       `json:"creationTime,omitempty"`
@@ -209,7 +212,7 @@ type PrivilegedPortalCapabilities struct {
 
 func GetByPolicyType(service *services.Service, policyType string) (*PolicySet, *http.Response, error) {
 	v := new(PolicySet)
-	relativeURL := fmt.Sprintf(mgmtConfigV1 + service.Client.Config.CustomerID + "/policySet/policyType/" + policyType)
+	relativeURL := mgmtConfigV1 + service.Client.Config.CustomerID + "/policySet/policyType/" + policyType
 	resp, err := service.Client.NewRequestDo("GET", relativeURL, common.Filter{MicroTenantID: service.MicroTenantID()}, nil, v)
 	if err != nil {
 		return nil, nil, err
@@ -242,6 +245,10 @@ func CreateRule(service *services.Service, rule *PolicyRule) (*PolicyRule, *http
 
 // PUT --> mgmtconfig​/v1​/admin​/customers​/{customerId}​/policySet​/{policySetId}​/rule​/{ruleId}
 func UpdateRule(service *services.Service, policySetID, ruleId string, policySetRule *PolicyRule) (*http.Response, error) {
+
+	ruleMutex.Lock()
+	defer ruleMutex.Unlock()
+
 	// Correct the initialization of Conditions slice with the correct type
 	if policySetRule != nil && len(policySetRule.Conditions) == 0 {
 		policySetRule.Conditions = []PolicyRuleResourceConditions{}
@@ -306,6 +313,9 @@ func GetByNameAndTypes(service *services.Service, policyTypes []string, ruleName
 
 // PUT --> /mgmtconfig/v1/admin/customers/{customerId}/policySet/{policySetId}/rule/{ruleId}/reorder/{newOrder}
 func Reorder(service *services.Service, policySetID, ruleId string, order int) (*http.Response, error) {
+	ruleMutex.Lock()
+	defer ruleMutex.Unlock()
+
 	path := fmt.Sprintf(mgmtConfigV1+service.Client.Config.CustomerID+"/policySet/%s/rule/%s/reorder/%d", policySetID, ruleId, order)
 	resp, err := service.Client.NewRequestDo("PUT", path, common.Filter{MicroTenantID: service.MicroTenantID()}, nil, nil)
 	if err != nil {
@@ -317,6 +327,9 @@ func Reorder(service *services.Service, policySetID, ruleId string, order int) (
 // PUT --> /mgmtconfig/v1/admin/customers/{customerId}/policySet/{policySet}/reorder
 // ruleIdOrders is a map[ruleID]Order
 func BulkReorder(service *services.Service, policySetType string, ruleIdToOrder map[string]int) (*http.Response, error) {
+	ruleMutex.Lock()
+	defer ruleMutex.Unlock()
+
 	policySet, resp, err := GetByPolicyType(service, policySetType)
 	if err != nil {
 		return resp, err