+++ zpa/services/policysetcontrollerv2/zpa_policy_access_redirection_rule_v2_test.go
@@ -22,7 +22,7 @@ func TestAccessRedirectionPolicyV2(t *testing.T) {
 	service := services.New(client)
 
 	// create service edge group for testing
-	svcEdgeGroup, _, err := serviceedgegroup.Create(service, serviceedgegroup.ServiceEdgeGroup{
+	svcEdgeGroup, _, err := serviceedgegroup.Create(service, &serviceedgegroup.ServiceEdgeGroup{
 		Name:                   name,
 		Description:            name,
 		Enabled:                true,