+++ zpa/services/serviceedgegroup/zpa_service_edge_group_test.go
@@ -19,8 +19,7 @@ func TestServiceEdgeGroup_Create(t *testing.T) {
 
 	service := services.New(client)
 
-	// Create new resource
-	createdResource, _, err := Create(service, ServiceEdgeGroup{
+	serviceEdgeGroup := ServiceEdgeGroup{
 		Name:                   name,
 		Description:            name,
 		Enabled:                true,
@@ -36,131 +35,142 @@ func TestServiceEdgeGroup_Create(t *testing.T) {
 		GraceDistanceEnabled:   true,
 		GraceDistanceValue:     "10",
 		GraceDistanceValueUnit: "MILES",
-	})
-	if err != nil {
-		t.Fatalf("Error creating resource: %v", err)
 	}
 
-	t.Run("TestResourceCreation", func(t *testing.T) {
-		if createdResource.ID == "" {
-			t.Error("Expected created resource ID to be non-empty, but got ''")
-		}
-		if createdResource.Name != name {
-			t.Errorf("Expected created resource name '%s', but got '%s'", name, createdResource.Name)
-		}
-	})
+	// Test resource creation
+	createdResource, _, err := Create(service, &serviceEdgeGroup)
 
-	t.Run("TestResourceRetrieval", func(t *testing.T) {
-		retrievedResource, _, err := Get(service, createdResource.ID)
-		if err != nil {
-			t.Fatalf("Error retrieving resource: %v", err)
-		}
-		if retrievedResource.ID != createdResource.ID {
-			t.Errorf("Expected retrieved resource ID '%s', but got '%s'", createdResource.ID, retrievedResource.ID)
-		}
-		if retrievedResource.Name != name {
-			t.Errorf("Expected retrieved resource name '%s', but got '%s'", name, createdResource.Name)
-		}
-	})
-
-	t.Run("TestResourceUpdate", func(t *testing.T) {
-		updatedResource := *createdResource
-		updatedResource.Name = updateName
-		_, err = Update(service, createdResource.ID, &updatedResource)
-		if err != nil {
-			t.Fatalf("Error updating resource: %v", err)
-		}
-	})
-
-	t.Run("TestResourceRetrievalByName", func(t *testing.T) {
-		retrievedResource, _, err := GetByName(service, updateName)
-		if err != nil {
-			t.Fatalf("Error retrieving resource by name: %v", err)
-		}
-		if retrievedResource.ID != createdResource.ID {
-			t.Errorf("Expected retrieved resource ID '%s', but got '%s'", createdResource.ID, retrievedResource.ID)
-		}
-		if retrievedResource.Name != updateName {
-			t.Errorf("Expected retrieved resource name '%s', but got '%s'", updateName, retrievedResource.Name)
-		}
-	})
-
-	t.Run("TestAllResourcesRetrieval", func(t *testing.T) {
-		resources, _, err := GetAll(service)
-		if err != nil {
-			t.Fatalf("Error retrieving groups: %v", err)
-		}
-		if len(resources) == 0 {
-			t.Error("Expected retrieved resources to be non-empty, but got empty slice")
-		}
-		found := false
-		for _, resource := range resources {
-			if resource.ID == createdResource.ID {
-				found = true
-				break
-			}
-		}
-		if !found {
-			t.Errorf("Expected retrieved groups to contain created resource '%s', but it didn't", createdResource.ID)
-		}
-	})
-
-	t.Run("TestResourceRemoval", func(t *testing.T) {
-		_, err := Delete(service, createdResource.ID)
-		if err != nil {
-			t.Fatalf("Error deleting resource: %v", err)
-		}
-	})
-}
-
-func TestRetrieveNonExistentResource(t *testing.T) {
-	client, err := tests.NewZpaClient()
+	// Check if the request was successful
 	if err != nil {
-		t.Fatalf("Error creating client: %v", err)
+		t.Errorf("Error making POST request: %v", err)
 	}
-	service := services.New(client)
 
-	_, _, err = Get(service, "non_existent_id")
+	if createdResource.ID == "" {
+		t.Error("Expected created resource ID to be non-empty, but got ''")
+	}
+	if createdResource.Name != name {
+		t.Errorf("Expected created resource name '%s', but got '%s'", name, createdResource.Name)
+	}
+	// Test resource retrieval
+	retrievedResource, _, err := Get(service, createdResource.ID)
+	if err != nil {
+		t.Errorf("Error retrieving resource: %v", err)
+	}
+	if retrievedResource.ID != createdResource.ID {
+		t.Errorf("Expected retrieved resource ID '%s', but got '%s'", createdResource.ID, retrievedResource.ID)
+	}
+	if retrievedResource.Name != name {
+		t.Errorf("Expected retrieved resource name '%s', but got '%s'", name, createdResource.Name)
+	}
+	// Test resource update
+	retrievedResource.Name = updateName
+	_, err = Update(service, createdResource.ID, retrievedResource)
+	if err != nil {
+		t.Errorf("Error updating resource: %v", err)
+	}
+	updatedResource, _, err := Get(service, createdResource.ID)
+	if err != nil {
+		t.Errorf("Error retrieving resource: %v", err)
+	}
+	if updatedResource.ID != createdResource.ID {
+		t.Errorf("Expected retrieved updated resource ID '%s', but got '%s'", createdResource.ID, updatedResource.ID)
+	}
+	if updatedResource.Name != updateName {
+		t.Errorf("Expected retrieved updated resource name '%s', but got '%s'", updateName, updatedResource.Name)
+	}
+
+	// Test resource retrieval by name
+	retrievedResource, _, err = GetByName(service, updateName)
+	if err != nil {
+		t.Errorf("Error retrieving resource by name: %v", err)
+	}
+	if retrievedResource.ID != createdResource.ID {
+		t.Errorf("Expected retrieved resource ID '%s', but got '%s'", createdResource.ID, retrievedResource.ID)
+	}
+	if retrievedResource.Name != updateName {
+		t.Errorf("Expected retrieved resource name '%s', but got '%s'", updateName, createdResource.Name)
+	}
+	// Test resources retrieval
+	resources, _, err := GetAll(service)
+	if err != nil {
+		t.Errorf("Error retrieving resources: %v", err)
+	}
+	if len(resources) == 0 {
+		t.Error("Expected retrieved resources to be non-empty, but got empty slice")
+	}
+	// check if the created resource is in the list
+	found := false
+	for _, resource := range resources {
+		if resource.ID == createdResource.ID {
+			found = true
+			break
+		}
+	}
+	if !found {
+		t.Errorf("Expected retrieved resources to contain created resource '%s', but it didn't", createdResource.ID)
+	}
+	// Test resource removal
+	_, err = Delete(service, createdResource.ID)
+	if err != nil {
+		t.Errorf("Error deleting resource: %v", err)
+		return
+	}
+
+	// Test resource retrieval after deletion
+	_, _, err = Get(service, createdResource.ID)
 	if err == nil {
-		t.Error("Expected error retrieving non-existent resource, but got nil")
+		t.Errorf("Expected error retrieving deleted resource, but got nil")
 	}
 }
 
-func TestDeleteNonExistentResource(t *testing.T) {
-	client, err := tests.NewZpaClient()
-	if err != nil {
-		t.Fatalf("Error creating client: %v", err)
-	}
-	service := services.New(client)
+// func TestRetrieveNonExistentResource(t *testing.T) {
+// 	client, err := tests.NewZpaClient()
+// 	if err != nil {
+// 		t.Fatalf("Error creating client: %v", err)
+// 	}
+// 	service := services.New(client)
 
-	_, err = Delete(service, "non_existent_id")
-	if err == nil {
-		t.Error("Expected error deleting non-existent resource, but got nil")
-	}
-}
+// 	_, _, err = Get(service, "non_existent_id")
+// 	if err == nil {
+// 		t.Error("Expected error retrieving non-existent resource, but got nil")
+// 	}
+// }
 
-func TestUpdateNonExistentResource(t *testing.T) {
-	client, err := tests.NewZpaClient()
-	if err != nil {
-		t.Fatalf("Error creating client: %v", err)
-	}
-	service := services.New(client)
+// func TestDeleteNonExistentResource(t *testing.T) {
+// 	client, err := tests.NewZpaClient()
+// 	if err != nil {
+// 		t.Fatalf("Error creating client: %v", err)
+// 	}
+// 	service := services.New(client)
 
-	_, err = Update(service, "non_existent_id", &ServiceEdgeGroup{})
-	if err == nil {
-		t.Error("Expected error updating non-existent resource, but got nil")
-	}
-}
+// 	_, err = Delete(service, "non_existent_id")
+// 	if err == nil {
+// 		t.Error("Expected error deleting non-existent resource, but got nil")
+// 	}
+// }
 
-func TestGetByNameNonExistentResource(t *testing.T) {
-	client, err := tests.NewZpaClient()
-	if err != nil {
-		t.Fatalf("Error creating client: %v", err)
-	}
-	service := services.New(client)
+// func TestUpdateNonExistentResource(t *testing.T) {
+// 	client, err := tests.NewZpaClient()
+// 	if err != nil {
+// 		t.Fatalf("Error creating client: %v", err)
+// 	}
+// 	service := services.New(client)
 
-	_, _, err = GetByName(service, "non_existent_name")
-	if err == nil {
-		t.Error("Expected error retrieving resource by non-existent name, but got nil")
-	}
-}
+// 	_, err = Update(service, "non_existent_id", &ServiceEdgeGroup{})
+// 	if err == nil {
+// 		t.Error("Expected error updating non-existent resource, but got nil")
+// 	}
+// }
+
+// func TestGetByNameNonExistentResource(t *testing.T) {
+// 	client, err := tests.NewZpaClient()
+// 	if err != nil {
+// 		t.Fatalf("Error creating client: %v", err)
+// 	}
+// 	service := services.New(client)
+
+// 	_, _, err = GetByName(service, "non_existent_name")
+// 	if err == nil {
+// 		t.Error("Expected error retrieving resource by non-existent name, but got nil")
+// 	}
+// }